# virtualBeacon_back/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'API velib'
  version: '0.1'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'api geo velib'
paths:
  '/places/delete/{id}':
    delete:
      operationId: 'App\Http\Controllers\ControllerPlaces::deletePosition'
      parameters:
        -
          name: id
          in: path
          description: 'Nom de la balise à supprimer'
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'Supression de la balise'
          content:
            application/json:
              schema:
                description: 'Tableau de la balise correspondante'
                type: array
                items:
                  $ref: '#/components/schemas/Suppression'
        '410':
          description: 'Erreur supression'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
  /places/create/position:
    post:
      operationId: 'App\Http\Controllers\ControllerPlaces::createPosition'
      responses:
        '200':
          description: 'Création nouvelle balise'
          content:
            application/json:
              schema:
                description: 'Tableau de la station correspondante'
                type: array
                items:
                  $ref: '#/components/schemas/Balises'
        '400':
          description: 'Erreur validité des données client'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
        '409':
          description: 'Nom de balise déja utilisé'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
  /places/position:
    post:
      operationId: 'App\Http\Controllers\ControllerPlaces::byPosition'
      responses:
        '200':
          description: 'La balise la plus proche'
          content:
            application/json:
              schema:
                description: 'Tableau de la station correspondante'
                type: array
                items:
                  $ref: '#/components/schemas/Balises'
        '400':
          description: 'Erreur validité des données client'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
  '/places/{id}':
    post:
      operationId: 'App\Http\Controllers\ControllerPlaces::byId'
      parameters:
        -
          name: id
          in: path
          description: 'Id de la balise recherchée'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Une balise appelée par son id'
          content:
            application/json:
              schema:
                description: 'Tableau de la balise correspondante'
                type: array
                items:
                  $ref: '#/components/schemas/Balises'
        '404':
          description: Erreur
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
  /places:
    get:
      operationId: 'App\Http\Controllers\ControllerPlaces::allPlaces'
      responses:
        '200':
          description: 'Toutes les balises enregistrées'
          content:
            application/json:
              schema:
                description: 'Tableau de toutes les balises enregistrées'
                type: array
                items:
                  $ref: '#/components/schemas/Balises'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                description: 'Tableau d''erreur'
                type: array
                items:
                  $ref: '#/components/schemas/Erreurs'
components:
  schemas:
    Balises:
      properties:
        id:
          type: integer
        lat:
          type: number
        lon:
          type: number
        name:
          type: string
        description:
          type: string
        user_id:
          type: integer
        created_at:
          type: string
        updtated_at:
          type: string
      type: object
    Erreurs:
      properties:
        name:
          type: string
        message:
          type: string
      type: object
    Suppression:
      properties:
        Message:
          type: string
        name:
          type: string
      type: object